+++ MHArmory.Search.Contracts/ISolver.cs
@@ -6,11 +6,11 @@ using System.Threading.Tasks;
 
 namespace MHArmory.Search.Contracts
 {
-    public interface ISolver
+    public interface ISolver : IDisposable
     {
         event Action<SearchMetrics> SearchMetricsChanged;
         event Action<double> SearchProgress;
 
-        Task<IList<ArmorSetSearchResult>> SearchArmorSets(CancellationToken cancellationToken);
+        Task<IList<ArmorSetSearchResult>> SearchArmorSets(ISolverData solverData, CancellationToken cancellationToken);
     }
 }